backend/requirements.txt
Flask
flask-cors
Pillow
requests
python-dotenv
onnxruntime